+++ TestWebsite/Client/Scripts/Library.d.ts
@@ -1,5 +1,4 @@
 ﻿declare module 'input-format'
-declare module 'react-select'
 declare module 'react-hint'
 
 declare const NProgress: any